Restructure store to track models by instance.alias
This uses a similar `_modelsFor` and `_fetchingFor` pattern found in
several other stores. The difference here is that we're not using the
BaseStore to avoid anything to deal with potential Liskov Substitution
principle violations [0]. Chief among them would be the _local_ cache.
The design intent here is that we don't want any of the convenience
associated with cached models; always fetch actions - just track what
you've fetched by :uuid (aka Instance alias).

[0] https://en.wikipedia.org/wiki/Liskov_substitution_principle